Auto-validate web clients -Steel
Right now, the permit system is really restrictive, so I've added in a hard-coded bypass for the system.

I do suggest we find another way of doing this that's more mod-and-player-friendly, maybe something interacting with Discord, as we have zero way of controlling who accesses AO through the web client now.